refactor: LegalInfo is moved to info.hpp
